% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{choose.vert}
\alias{choose.vert}
\title{Helper for choosing vertices with the min/max l2 norm.}
\usage{
choose.vert(V, type = 1)
}
\description{
Helper for choosing vertices with the min/max l2 norm.
}
\keyword{internal}
