% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dep.assign.rank}
\alias{dep.assign.rank}
\title{Helper for Assigning Ranks for Test of Independence}
\usage{
dep.assign.rank(XY, elem, rank.id)
}
\description{
Helper for Assigning Ranks for Test of Independence
}
\keyword{internal}
