% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dep1D}
\alias{dep1D}
\title{1D Test of Independece Helper}
\usage{
dep1D(XY, center, epsilon, maxit, verbose)
}
\arguments{
\item{XY}{input data matrix.}

\item{center}{logical indicating if the centroids should be computed.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating wether to display optimization messages.}
}
\value{
a matrix, represents either the centroid or the cells of the combined optimal transport map.
}
\description{
Compute the quantiles of U with respect to (X, Y).
}
\keyword{internal}
