% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dep2D}
\alias{dep2D}
\title{Helper for computing the 1D test of independece}
\usage{
dep2D(XY, U, center, epsilon, maxit, verbose)
}
\arguments{
\item{XY}{input data matrix.}

\item{U}{sequence used to evaluate the integral.}

\item{center}{logical indicating if the centroids should be computed.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating wether to display optimization messages.}
}
\value{
a list, which contains the quantile indices, and the vertices of the combined optimal transport map.
}
\description{
Compute the quantiles of U with respect to (X, Y).
}
\keyword{internal}
