% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth.uniform}
\alias{depth.uniform}
\title{Multivariate Uniform Depth}
\usage{
depth.uniform(X)
}
\arguments{
\item{X}{a \eqn{n} by \eqn{d} numeric matrix.}
}
\value{
a vector containing depth of the data.
}
\description{
Compute the depth associated with U[0, 1]^d.
}
