% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dualGraphs2D}
\alias{dualGraphs2D}
\title{2D Semi-discrete Optimal Transport Map}
\usage{
dualGraphs2D(X, epsilon, maxit, verbose)
}
\arguments{
\item{X}{input data matrix.}

\item{epsilon}{maximum error for the optimization algorithm.}

\item{maxit}{maximum number of solver iterations.}

\item{verbose}{logical indicating whether to display optimization messages.}
}
\value{
a list describing the resulting optimal transport map.
}
\description{
2D Semi-discrete Optimal Transport Map
}
\keyword{internal}
