% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract.segment}
\alias{extract.segment}
\title{Helper for extracting line segments for visualization}
\usage{
extract.segment(V, id_col, vert_col)
}
\description{
Helper for extracting line segments for visualization
}
\keyword{internal}
