% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gof.assign.rank}
\alias{gof.assign.rank}
\title{Helper for Assigning Ranks for Goodness-of-fit Test}
\usage{
gof.assign.rank(elem, rank.id, d)
}
\description{
Helper for Assigning Ranks for Goodness-of-fit Test
}
\keyword{internal}
