% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gof3DHelper}
\alias{gof3DHelper}
\title{3D Goodness-of-fit Test Helper}
\usage{
gof3DHelper(X, Y, U, epsilon, maxit, verbose)
}
\arguments{
\item{X}{input data matrix.}

\item{Y}{input data matrix.}

\item{U}{sequence used to evaluate the integral.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating whether to display optimization messages.}
}
\value{
a list containing the quantile indices for \eqn{X} and \eqn{Y}.
}
\description{
Compute the semi-discrete optimal transport quantiles of quasi-MC sequence \eqn{U} with respect to \eqn{X} and \eqn{Y}.
}
\keyword{internal}
