% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoF_2D.R
\name{GoF_2D}
\alias{GoF_2D}
\title{2D goodness-of-test statistic}
\usage{
GoF_2D(X, Y, mc = 1000, type = "max", epsilon = 0.001, maxit = 100,
  na.rm = F)
}
\arguments{
\item{X}{Sample data matrix, of size \eqn{n} by \eqn{2}.}

\item{Y}{Sample data matrix, of size \eqn{m} by \eqn{2}.}

\item{mc}{Number of Monte Carlo iteration samples used to evaluate the test statistic.}

\item{type}{Method used to choose the rank over the Voronoi cell boundary.}

\item{maxit}{Max number of iterations before termination.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}

\item{verbose}{Wether to display messages during optimization.}
}
\description{
Computes the 2D goodness-of-test statistic using rank defined through the semi-continuous optimal transport map.
}
