% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jointRankHelper3D}
\alias{jointRankHelper3D}
\title{3D Joint Samples Rank Helper}
\usage{
jointRankHelper3D(XY, center, epsilon, maxit, verbose)
}
\arguments{
\item{XY}{3D input data matrix.}

\item{center}{logical indicating if the centroids should be computed.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating whether to display optimization messages.}
}
\value{
a matrix, represents either the centroids/cells of the RVD of joint samples.
}
\description{
Helps computing the Empirical Rank for 3D Joint Samples \eqn{(X, Y)}.
}
\keyword{internal}
