% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{locateRDT2D}
\alias{locateRDT2D}
\title{Locate Points within a 2D RDT}
\usage{
locateRDT2D(Q, V)
}
\arguments{
\item{Q}{input query matrix.}

\item{V}{matrix of sorted triangle vertices.}
}
\value{
a vector of indices indicating which triangles the query points belong.
}
\description{
Find the RDT triangles where a set Q of query points belongs to.
}
\keyword{internal}
