% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{locateRVD2D}
\alias{locateRVD2D}
\title{Locate within a 2D RVD}
\usage{
locateRVD2D(Q, X, w)
}
\arguments{
\item{Q}{input query matrix.}

\item{X}{input data used to build the RVD.}

\item{w}{weights of the RVD cells.}
}
\value{
a vector of indices indicating where the query points are transported to.
}
\description{
Find the RVD cells where a set Q of query points is transported to.
}
\keyword{internal}
