% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R
\name{otm.dep.test}
\alias{otm.dep.test}
\title{1D Test of Independence}
\usage{
otm.dep.test(X, Y, scale = c(0, 1), rank.data = "uniform",
  epsilon = 1e-06, maxit = 100, verbose = F, na.rm = F)
}
\arguments{
\item{X}{input data vector.}

\item{Y}{input data vector.}

\item{scale}{a numeric vector indicating the minimum and maximum of the scaled data. Set to \code{NULL} to skip scaling.}

\item{rank.data}{choose the method for assigning ranks to the data points. 
Can be "\code{max}", "\code{min}", "\code{center}", or "\code{uniform}".}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating whether to display optimization messages.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{mc}{number of quasi-Monte-Carlo samples used to evaluate the test statistic.}
}
\value{
the value of the  mutual independence test statistic.
}
\description{
\code{otm.dep.test} computes the 1D mutual independence test statistic using ranks defined through the semi-discrete optimal transport map.
}
\details{
Given samples \eqn{(X_1, Y_1), \dots, (X_n, Y_n)}, we use the following statistic for test of independence:
\deqn{T_n=\sum_{i=1}^n\|\hat{R}(X_i, Y_i)-\tilde{R}(X_i, Y_i)\|^2.}
}
\seealso{
\code{\link{otm.rank}} for optimal transport rank.
}
\keyword{htest,}
\keyword{multivariate}
