% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{otm.depth.OTM_2D}
\alias{otm.depth.OTM_2D}
\title{2D Semi-discrete Optimal Transport Depth}
\usage{
\method{otm.depth}{OTM_2D}(object, X)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{X}{a \eqn{n} by \eqn{2} numeric matrix.}
}
\value{
a vector containing depth of the data.
}
\description{
Compute the MK depth relative to the 2D standard uniform distribution.
}
