% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{otm.depth}
\alias{otm.depth}
\title{Semi-discrete Optimal Transport Depth}
\usage{
otm.depth(object, X, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{X}{a \eqn{n} by \eqn{2} numeric matrix.}
}
\value{
a vector containing depth of the data.
}
\description{
Compute the MK depth relative to the multivariate uniform distribution U[0, 1]^d.
}
