% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otm.R
\name{otm.fit}
\alias{otm.fit}
\title{Uniform Semi-discrete Optimal Transport Map}
\usage{
otm.fit(data, epsilon = 0.001, maxit = 100, verbose = F, na.rm = F)
}
\arguments{
\item{data}{input data matrix, of size \eqn{n} by \eqn{d}.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating wether to display optimization messages.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
\code{otm.fit} returns an object of class "\code{otm.2d}" or "\code{otm.3d}", depending on the dimesion of input data.
An object of class "\code{otm}" is a list describing the resulting optimal transport map.
}
\description{
Computes the semi-discrete optimal transport map from U[0, 1]^d measure to the input data set.
}
\keyword{graphs}
\keyword{optimize,}
