% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{otm.potential.OTM.2D}
\alias{otm.potential.OTM.2D}
\title{2D Semi-discrete Optimal Transport Dual Potential}
\usage{
\method{otm.potential}{OTM.2D}(object, X)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{X}{a \eqn{n} by \eqn{2} numeric matrix.}
}
\value{
a matrix containing dual potentials of the data.
}
\description{
Compute the Legendre-Fenchel dual of the Alexandrov potential.
}
