% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{otm.quantile}
\alias{otm.quantile}
\title{Semi-discrete Optimal Transport Quantile}
\usage{
otm.quantile(object, Q, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}
}
\value{
a list containing the quantiles of the data, and the indices of the corresponding cells.
}
\description{
Compute the quantiles based on optimal transport.
}
