% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{otm.rank.otm.2d}
\alias{otm.rank.otm.2d}
\title{2D Semi-discrete Optimal Transport Rank}
\usage{
\method{otm.rank}{otm.2d}(object, Q, use.geo = FALSE, ...)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}

\item{use.geo}{logical indicating if the geometric method should be used to compute the ranks.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a list containing the ranks of the data and the corresponding convex conjugate potential values.
If \code{use.geo == TRUE} then the convex conjugates will not be computed (\code{NA}s will be returned).
}
\description{
The 2D implementation of \code{otm.rank}.
}
\keyword{internal}
