% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{otm.rank.OTM_2D}
\alias{otm.rank.OTM_2D}
\title{2D Semi-discrete Optimal Transport Rank}
\usage{
\method{otm.rank}{OTM_2D}(object, X, h = 1e-07)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{X}{a \eqn{n} by \eqn{2} numeric matrix.}

\item{h}{step size for the numerical differentiation.}
}
\value{
a matrix containing rank of the data.
}
\description{
Compute the rank function based on optimal transport.
}
