% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{otm.rank}
\alias{otm.rank}
\title{Semi-discrete Optimal Transport Rank}
\usage{
otm.rank(object, Q, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}
}
\value{
a matrix containing the ranks of the data.
}
\description{
Compute the ranks based on optimal transport.
}
