% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTM_2D.R
\name{OTM_2D}
\alias{OTM_2D}
\title{2D semi-continuous optimal transport map}
\usage{
OTM_2D(data, epsilon = 0.001, maxit = 100, verbose = F, na.rm = F)
}
\arguments{
\item{data}{Input coordinate matrix, of size \eqn{n} by \eqn{2}.}

\item{epsilon}{Convergence threshold for optimization.}

\item{maxit}{Max number of iterations before termination.}

\item{verbose}{Wether to display messages during optimization.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
A list describing the resulting Laguerre diagram.
}
\description{
Computes the semi-continuous optimal transport map from a uniform measure on [0, 1]^2 to a given data set.
}
\keyword{Laguerre}
\keyword{Voronoi,}
