% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.OTM_2D.R
\name{plot.OTM_2D}
\alias{plot.OTM_2D}
\title{2D semi-continuous optimal transport map plot}
\usage{
\method{plot}{OTM_2D}(Obj, col.data = "cornflowerblue",
  col.center = "firebrick", col.edge = "black", connect = F, ...)
}
\arguments{
\item{Obj}{Fitted optimal transport map.}
}
\description{
Plot OTM
}
\keyword{Laguerre}
\keyword{Voronoi,}
