% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otm.R
\name{plot.tos.2d}
\alias{plot.tos.2d}
\title{Plotting 2D Semi-discrete Optimal Transport Map}
\usage{
\method{plot}{tos.2d}(
  x,
  which = "Both",
  col.data = "dodgerblue3",
  col.center = "darkorange1",
  col.edge = "black",
  draw.data = TRUE,
  draw.center = TRUE,
  draw.map = FALSE,
  xlim = c(0, 1),
  ylim = c(0, 1),
  xlab = expression("x"),
  ylab = expression("y"),
  pch = 20,
  ...
)
}
\arguments{
\item{x}{a fitted \code{tos.2d} object.}

\item{which}{specify which graph(s) to plot. Can be \code{None}, \code{RVD}, \code{RDT}, or \code{Both}.}

\item{col.data}{color of the data points.}

\item{col.center}{color of the Voronoi centroids.}

\item{col.edge}{color of the edges in plotting RVD and RDT.}

\item{draw.data}{logical indicating if the data points should be plotted.}

\item{draw.center}{logical indicating if the centroids should be plotted.}

\item{draw.map}{logical indicating if dashed lines should be added to show mapping between the data and the Voronoi cells.}

\item{xlim}{the x limits of the plot.}

\item{ylim}{the y limits of the plot.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{pch}{a vector of plotting characters or symbols.}

\item{\dots}{other graphical parameters to plot.}
}
\description{
\code{plot.tos.2d} plots the restricted Voronoi diagram (RVD) and the restricted Delaunay triangulation (RDT) of a given
2D semi-discrete optimal transport map.
}
\examples{
# generate some data
X = c(0.5, 0.8, -0.2, -1.5, 1.4, 0.5, -1.1, -0.1, -1.1, -2.6)
Y = c(1.6, -1.0, -0.1, 0.5, -1.3, 2.9, -0.4, 1.3, -1.8, -2.5)

# compute the optimal transport map from U[0, 1]^2 to the data
XY.OTM = tos.fit(cbind(X, Y))

# plot the RVD
plot(XY.OTM, which = "RVD", draw.map = TRUE)
}
\keyword{hplot}
