% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otm.R
\name{plot.tos.3d}
\alias{plot.tos.3d}
\title{Plot the 3D Semi-discrete Optimal Transport Map}
\usage{
\method{plot}{tos.3d}(x, interactive = TRUE, which = "Both",
  col.data = "dodgerblue3", col.center = "darkorange1",
  col.rvd = "black", col.rdt = "firebrick3", draw.data = T,
  draw.center = T, draw.map = F, draw.id = NULL, xlim = c(0, 1),
  ylim = c(0, 1), zlim = c(0, 1), xlab = expression("x"),
  ylab = expression("y"), zlab = expression("z"), size = 10,
  phi = 45, theta = 45, pch = 20, ...)
}
\arguments{
\item{x}{a fitted \code{tos.2d} object.}

\item{interactive}{logical indicating if the plot should be interactive.}

\item{which}{specify which graph(s) to plot. Can be \code{None}, \code{RVD}, \code{RDT}, or \code{Both}.}

\item{col.data}{color of the data points.}

\item{col.center}{color of the Voronoi centroids.}

\item{col.rvd}{color of the edges in RVD plot.}

\item{col.rdt}{color of the edges in RDT plot.}

\item{draw.data}{logical indicating if the data points should be plotted.}

\item{draw.center}{logical indicating if the centroids should be plotted.}

\item{draw.map}{logical indicating if dashed lines should be added to show mapping between the data and the Voronoi cells.}

\item{draw.id}{vector of indices specifying which data points (and the corresponding cells in RVD and RDT) will be plotted. 
Set it to \code{NULL} to plot all points. When it is not \code{NULL} and RDT is to be drawn, 
\code{plot.tos.3d} will plot all tetrahedra that have points in \code{draw.id} as one of its vertices.}

\item{xlim}{the x limits of the plot.}

\item{ylim}{the y limits of the plot.}

\item{zlim}{the z limits of the plot.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{zlab}{a label for the z axis.}

\item{size}{size of the plotted points in the interactive plots.}

\item{theta, phi}{the angles defining the viewing direction for non-interactive plots. \code{theta} gives the azimuthal direction and \code{phi} the colatitude.}

\item{pch}{a vector of plotting characters or symbols.}

\item{\dots}{other graphical parameters to plot.}
}
\description{
Plots the restricted Voronoi diagram (RVD) and the restricted Delaunay triangulation (RDT) of a given
3D semi-discrete optimal transport map.
}
\keyword{hplot}
