% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.OTM_2D.R
\name{potential.OTM_2D}
\alias{potential.OTM_2D}
\title{2D semi-continuous optimal transport dual potential}
\usage{
potential.OTM_2D(object, X)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{x}{a \eqn{n} by \eqn{2} numeric matrix.}
}
\value{
a matrix containing dual potentials of the data.
}
\description{
Compute the Legendre-Fenchel dual of the Alexandrov potential
2D semi-discrete optimal transport dual potential.
}
\keyword{Laguerre}
\keyword{Voronoi,}
