% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rand.perm}
\alias{rand.perm}
\title{Generate \eqn{m} Distinct Permutations of \eqn{n} Elements Uniformly.}
\usage{
rand.perm(n, m)
}
\description{
Generate \eqn{m} Distinct Permutations of \eqn{n} Elements Uniformly.
}
\keyword{internal}
