% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scaling.min.max}
\alias{scaling.min.max}
\title{Min-Max scaling}
\usage{
scaling.min.max(data, min = 0, max = 1)
}
\arguments{
\item{data}{a numeric matrix for the input data.}

\item{min}{desired minimum of the transformed data.}

\item{max}{desired maximum of the transformed data.}
}
\value{
a matrix containing the scaled data.
}
\description{
\code{scaling.min.max} rescales the range of a data set to [min, max].
}
\examples{
# generate some data
X = c(1.5, 3.2, 2.6, 0.3, 5.1)
Y = c(3.3, 6.4, 1.8, 0.7, 2.2)

# scale (X, Y) into [0, 1] range
scaling.min.max(cbind(X, Y), 0, 1)
}
\keyword{utilities}
