% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testOTM-package.R
\docType{package}
\name{testOTM-package}
\alias{testOTM}
\alias{testOTM-package}
\title{testOTM: Multivariate Ranks and Quantiles using Optimal Transportation}
\description{
This package implements the multivariate ranks and quantiles, together with their application in statistical testing, that based on the semi-discrete optimal transportation. The underlying geometric computation is performed by the Geogram library.
}
\references{
Geogram, \url{http://alice.loria.fr/software/geogram/doc/html/index.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Francis-Hsu/testOTM}
  \item Report bugs at \url{https://github.com/Francis-Hsu/testOTM/issues}
}

}
\author{
\strong{Maintainer}: Peng Xu \email{px2132@columbia.edu}

}
\keyword{internal}
