% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testOTM-package.R
\docType{package}
\encoding{UTF-8}
\name{testOTM}
\alias{testOTM}
\alias{testOTM-package}
\title{testOTM: Multivariate Ranks and Quantiles by Optimal Transport}
\description{
Multivariate ranks and quantiles based on the theory of optimal transportation, with applications in non-parametric testing.
}
\references{
\itemize{
\item Geogram, \url{http://alice.loria.fr/software/geogram/doc/html/index.html}.
\item Ghosal, Promit, and Bodhisattva Sen. 2019. “Multivariate Ranks and Quantiles Using Optimal Transportation and Applications to Goodness-of-Fit Testing.” \url{http://arxiv.org/abs/1905.05340}.
\item Lévy, Bruno, and Erica L. Schwindt. 2018. “Notions of Optimal Transport Theory and How to Implement Them on a Computer.” Computers & Graphics 72: 135–48. \url{https://doi.org/10.1016/j.cag.2018.01.009}.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Francis-Hsu/testOTM}
  \item Report bugs at \url{https://github.com/Francis-Hsu/testOTM/issues}
}

}
\author{
\strong{Maintainer}: Peng Xu \email{px2132@columbia.edu}

Other contributors:
\itemize{
  \item Bruno Lévy (author of Geogram) [contributor, copyright holder]
}

}
\keyword{internal}
