% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R
\name{tos.dep.test}
\alias{tos.dep.test}
\title{1D (Permutation) Test of Independence}
\usage{
tos.dep.test(
  X,
  Y,
  n.perm = 0,
  rank.data = "uniform",
  epsilon = 1e-06,
  maxit = 100,
  verbose = FALSE
)
}
\arguments{
\item{X}{input data vector.}

\item{Y}{input data vector.}

\item{n.perm}{number of permutations used for computing \eqn{p}-value.}

\item{rank.data}{choose the method for assigning ranks to the data points. 
Can be \code{max}, \code{min}, \code{center}, or \code{uniform}.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating whether to display optimization messages.}
}
\value{
a list which contains the permutation test statistics and the \eqn{p}-value.
}
\description{
\code{tos.dep.test} computes the 1D mutual independence test statistic using ranks defined through 
the semi-discrete optimal transport map.
}
\details{
\code{tos.dep.test} tests the null hypothesis that \eqn{X} and \eqn{Y} are independent. 
The \eqn{p}-value is computed through permutations. For a very small sample (size less than \eqn{8}), 
all possible permutations are generated (assuming a suitable \code{n.perm} is provided). 
For larger sample size, Monte Carlo permutation sampling is used to approximate the permutation \eqn{p}-value.

Given samples \eqn{Z_1=(X_1, Y_1), \dots, Z_n=(X_n, Y_n)}, the following statistic is used for the test of independence:
\deqn{T_n=\sum_{i=1}^n\|\hat{R}(Z_i)-\tilde{R}(Z_i)\|_2^2,} 
where \eqn{\tilde{R}} is the joining of the usual 1D ranks.
}
\examples{
# genrate a sample
X = rnorm(50)

# generate another sample
Y = rnorm(50)

# test whether the two samples are independent
tos.dep.test(X, Y, n.perm = 50)
}
\references{
Promit Ghosal and Bodhisattva Sen (2019). 
\emph{Multivariate Ranks and Quantiles Using Optimal Transportation and Applications to Goodness-of-Fit Testing}.
\url{http://arxiv.org/abs/1905.05340}.
}
\seealso{
\code{\link{tos.rank}} for semi-discrete optimal transport rank.
}
\keyword{htest}
\keyword{multivariate}
