% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{tos.depth.tos.3d}
\alias{tos.depth.tos.3d}
\title{3D Semi-discrete Optimal Transport Depth}
\usage{
\method{tos.depth}{tos.3d}(object, query, scale = TRUE, rank.data = "center", ...)
}
\arguments{
\item{object}{a fitted 3D optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point.}

\item{rank.data}{choose the method for assigning ranks to the data points.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a vector containing the depths of the queries.
}
\description{
The 3D implementation of \code{tos.depth}.
}
\keyword{internal}
