% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R
\name{tos.gof.test}
\alias{tos.gof.test}
\title{2D/3D Goodness-of-fit Test}
\usage{
tos.gof.test(
  X,
  Y,
  mc = 10000,
  n.perm = 0,
  scale = c(0, 1),
  rank.data = "uniform",
  epsilon = 1e-06,
  maxit = 100,
  verbose = F,
  na.rm = F
)
}
\arguments{
\item{X}{input data matrix, of size \eqn{n} by \eqn{2} or \eqn{3}.}

\item{Y}{input data matrix, of size \eqn{m} by \eqn{2} or \eqn{3}.}

\item{mc}{number of quasi-Monte Carlo samples used to evaluate the test statistic.}

\item{n.perm}{number of permutations used for computing \eqn{p}-value.}

\item{scale}{a numeric vector indicating the minimum and maximum of the scaled data. Set to \code{NULL} to skip scaling.}

\item{rank.data}{choose the method for assigning ranks to the data points. 
Can be \code{max}, \code{min}, \code{center}, or \code{uniform}. 
\code{uniform} is not currently implemented for 3D, and \code{center} will be used if it is chosen.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating whether to display optimization messages.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
a list which contains the permutation test statistics and the \eqn{p}-value.
}
\description{
\code{tos.gof.test} computes the 2D/3D goodness-of-test statistic using ranks defined through the semi-discrete optimal transport map.
}
\details{
\code{tos.gof.test} tests whether the samples come from the same population. 
The \eqn{p}-value is computed through permutations. For a very small sample (size less than \eqn{8}), 
all possible permutations are generated (assuming a suitable \code{n.perm} is provided). 
For larger sample size, Monte Carlo permutation sampling is used to approximate the permutations \eqn{p}-value.

Given samples \eqn{X_1, \dots, X_n} and \eqn{Y_1, \dots, Y_m}, we use the following statistic for goodness-of-fit testing:
\deqn{T_{X,Y}=\int_{[0, 1]^d}\|\hat{R}_{X,Y}[\hat{Q}_{X}(u)]-\hat{R}_{X,Y}[\hat{Q}_{Y}(u)]\|_2^2\,d\mu(u),} 
where \eqn{\mu\sim U[0, 1]^d}. Evaluation of this integral is done through quasi-Monte Carlo with Sobol sequence.
}
\references{
Promit Ghosal and Bodhisattva Sen. 2019. 
\emph{Multivariate Ranks and Quantiles Using Optimal Transportation and Applications to Goodness-of-Fit Testing}.
\url{http://arxiv.org/abs/1905.05340}.
}
\seealso{
\code{\link{tos.rank}} for optimal transport rank.
}
\keyword{htest}
\keyword{multivariate}
