% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{tos.potential.tos.2d}
\alias{tos.potential.tos.2d}
\title{2D Semi-discrete Optimal Transport Potential}
\usage{
\method{tos.potential}{tos.2d}(object, query, scale = TRUE, ...)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point
(or a vector of length 2 representing a single point).}

\item{scale}{logical indicating if the queries should be scaled.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a matrix containing the potentials of the queries.
}
\description{
The 2D implementation of \code{tos.potential}.
}
\examples{
# generate some data
X = c(0.5, 0.8, -0.2, -1.5, 1.4, 0.5, -1.1, -0.1, -1.1, -2.6)
Y = c(1.6, -1.0, -0.1, 0.5, -1.3, 2.9, -0.4, 1.3, -1.8, -2.5)

# compute the optimal transport map from U[0, 1]^2 to the data
XY.OTM = tos.fit(cbind(X, Y))

# compute the potential of point (0, 0)
tos.potential(XY.OTM, c(0, 0))
}
\keyword{internal}
