% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{tos.quantile}
\alias{tos.quantile}
\title{Semi-discrete Optimal Transport Quantiles}
\usage{
tos.quantile(object, query, scale = TRUE, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point
(or a vector of length 2 or 3 representing a single point).}

\item{scale}{logical indicating if the queries should be scaled.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a list containing the quantiles of the queries and the indices of the corresponding cells.
}
\description{
\code{tos.quantile} computes the optimal transport quantiles.
}
\examples{
# generate some data
X = c(0.5, 0.8, -0.2, -1.5, 1.4, 0.5, -1.1, -0.1, -1.1, -2.6)
Y = c(1.6, -1.0, -0.1, 0.5, -1.3, 2.9, -0.4, 1.3, -1.8, -2.5)

# compute the optimal transport map from U[0, 1]^2 to the data
XY.OTM = tos.fit(cbind(X, Y))

# compute the quantile of point (0, 0).
tos.quantile(XY.OTM, c(0, 0))
}
\references{
Promit Ghosal and Bodhisattva Sen (2019). 
\emph{Multivariate Ranks and Quantiles Using Optimal Transportation and Applications to Goodness-of-Fit Testing}.
\url{http://arxiv.org/abs/1905.05340}.
}
\keyword{multivariate}
