% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{tos.rank}
\alias{tos.rank}
\title{Semi-discrete Optimal Transport Rank}
\usage{
tos.rank(
  object,
  query,
  scale = TRUE,
  rank.data = "uniform",
  rank.algo = "lp",
  ...
)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point 
(or a vector of length 2 or 3 representing a single point).}

\item{scale}{logical indicating if the queries should be scaled.}

\item{rank.data}{choose the method for assigning ranks to the data points. 
Can be \code{max}, \code{min}, \code{center}, or \code{uniform}.}

\item{rank.algo}{choose the algorithm for computing the ranks. Can be \code{lp} or \code{geom}.}

\item{\dots}{additional arguments.}
}
\value{
a list containing the ranks of the data and the corresponding convex conjugate potential values.
If \code{rank.algo = "geom"} or data points are presented in the queries,
then the corresponding conjugate potentials will not be computed (\code{NA}s will be returned).
}
\description{
\code{tos.rank} computes the optimal transport ranks.
}
\details{
The rank of the a data points used to fit the optimal transport map is not uniquely defined. 
In principle it can be any point within the corresponding Laguerre cell. 
Set \code{rank.data} to \code{max} to use the vertices with the maximum l2 norm (\code{min} for minimum l2 norm),
which is analogous to the usual ECDF rank in 1D. Option \code{center} will assign the centroid of the cell as the rank.
Finally, option \code{uniform} will sample a point uniformly from the Laguerre cell as the rank.

The rank for a non-data query point can be computed by solving a series of linear programs, 
or by finding the intersection of Laguerre cells that correspond to the simplex containing the query.
Set \code{rank.algo} to \code{lp} to use the former and \code{geom} for the latter.
Note the geometric method is slow and does not work for points lie outside the RDT, it is also not implemented for 3D, 
\code{lp} method will be used in these cases.
}
\examples{
# generate some data
X = c(0.5, 0.8, -0.2, -1.5, 1.4, 0.5, -1.1, -0.1, -1.1, -2.6)
Y = c(1.6, -1.0, -0.1, 0.5, -1.3, 2.9, -0.4, 1.3, -1.8, -2.5)

# compute the optimal transport map from U[0, 1]^2 to the data
XY.OTM = tos.fit(cbind(X, Y))

# compute the rank of point (0, 0).
tos.rank(XY.OTM, c(0, 0))
}
\references{
Promit Ghosal and Bodhisattva Sen (2019). 
\emph{Multivariate Ranks and Quantiles Using Optimal Transportation and Applications to Goodness-of-Fit Testing}.
\url{http://arxiv.org/abs/1905.05340}.
}
\seealso{
\code{\link{tos.gof.test}} and \code{\link{tos.dep.test}} for non-parametric tests using optimal transport rank.
}
\keyword{multivariate}
