% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{tos.rank.tos.3d}
\alias{tos.rank.tos.3d}
\title{3D Semi-discrete Optimal Transport Rank}
\usage{
\method{tos.rank}{tos.3d}(object, query, scale = TRUE, rank.data = "center", ...)
}
\arguments{
\item{object}{a fitted 3D optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point
(or a vector of length 3 representing a single point).}

\item{scale}{logical indicating if the queries should be scaled.}

\item{rank.data}{choose the method for assigning ranks to the data points. 
Can be \code{max}, \code{min}, or \code{center}.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a list containing the ranks of the data and the corresponding convex conjugate potential values.
}
\description{
The 3D implementation of \code{tos.rank}.
}
\examples{
# generate some data
X = c(2.3, 0.8, 1.2, 0.6, 0.2, 0.5, -0.7, -0.6, -0.2, 3.2)
Y = c(1.5, -0.8, 0.6, -2.1, 1.3, 1.3, -1.1, 1.0, -0.2, 1.0)
Z = c(0.8, -0.3, -1.3, -1.8, 1.2, -0.7, -0.8, 1.3, -0.9, 2.0)

# compute the optimal transport map from U[0, 1]^3 to the data
XYZ.OTM = tos.fit(cbind(X, Y, Z))

# compute the rank of point (0, 0, 0).
tos.rank(XYZ.OTM, c(0, 0, 0))
}
\keyword{internal}
