% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unscaling.min.max}
\alias{unscaling.min.max}
\title{Undo min-max scaling}
\usage{
unscaling.min.max(data, center, scale)
}
\arguments{
\item{data}{a numeric matrix for the input data.}

\item{center}{a numeric vector containing the locations for transforming the data.}

\item{scale}{a numeric vector containing the scales for transforming the data.}
}
\value{
a matrix containing the unscaled data.
}
\description{
Undo min-max scaling
}
\keyword{utilities}
