% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{choose.vert}
\alias{choose.vert}
\title{Helper for Choosing Vertices with the Min/Max \eqn{l_2} Norm.}
\usage{
choose.vert(V, type = 1)
}
\description{
Helper for Choosing Vertices with the Min/Max \eqn{l_2} Norm.
}
\keyword{internal}
