% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth.uniform}
\alias{depth.uniform}
\title{Multivariate Uniform Depth Function}
\usage{
depth.uniform(query)
}
\arguments{
\item{query}{a numeric matrix where each row represents a query point.}
}
\value{
a vector containing the depths of the queries.
}
\description{
\code{depth.uniform} computes the depths associated with the \eqn{U[0, 1]^d} measure.
}
\keyword{internal}
