% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{depth.uniform}
\alias{depth.uniform}
\title{Multivariate Uniform Depth Function}
\usage{
depth.uniform(Q)
}
\arguments{
\item{Q}{a numeric matrix where each row represents a query point.}
}
\value{
a vector containing the depths of the data.
}
\description{
Compute the depths associated with U[0, 1]^d.
}
