% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gof2D}
\alias{gof2D}
\title{Helper for computing the goodness-of-fit test statistics}
\usage{
gof2D(X, Y, U, center, epsilon, maxit, verbose)
}
\arguments{
\item{X}{input data matrix.}

\item{Y}{input data matrix.}

\item{U}{sequence used to evaluate the integral.}

\item{center}{logical indicating if the centroids should be computed.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating wether to display optimization messages.}
}
\value{
a list, which contains the quantile indices, and the vertices of the combined optimal transport map.
}
\description{
Compute the quantiles of U with respect to X and Y, as well as the optimal transport map of the combined data.
}
\keyword{internal}
