% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoF_2D.R
\name{GoF_2D}
\alias{GoF_2D}
\title{2D Goodness-of-test Statistic}
\usage{
GoF_2D(X, Y, mc = 1000, type = "max", epsilon = 0.001, maxit = 100,
  verbose = F, na.rm = F)
}
\arguments{
\item{X}{sample data matrix, of size \eqn{n} by \eqn{2}.}

\item{Y}{sample data matrix, of size \eqn{m} by \eqn{2}.}

\item{mc}{number of Monte Carlo iteration samples used to evaluate the test statistic.}

\item{type}{method used to choose the rank over the Voronoi cell boundary.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating wether to display optimization messages.}

\item{na.rm}{logical indicating whether NA values should be stripped before the computation proceeds.}
}
\description{
Computes the 2D goodness-of-test statistic using rank defined through the semi-discrete optimal transport map.
}
\keyword{htest}
\keyword{multivariate,}
