% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest.R
\name{otm.dep.test}
\alias{otm.dep.test}
\title{1D Test of Independence}
\usage{
otm.dep.test(X, Y, rank = "center", epsilon = 0.001, maxit = 100,
  verbose = F, na.rm = F)
}
\arguments{
\item{X}{input data vector.}

\item{Y}{input data vector.}

\item{rank}{choose the method for assigning ranks to the data points. 
Can be "\code{max}", "\code{min}", "\code{center}", or "\code{uniform}".}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating wether to display optimization messages.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{mc}{number of quasi-Monte-Carlo samples used to evaluate the test statistic.}
}
\value{
the value of the  mutual independence test statistic.
}
\description{
\code{otm.dep.test} computes the 1D mutual independence test statistic using ranks defined through the semi-discrete optimal transport map.
}
\keyword{htest,}
\keyword{multivariate}
