% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{otm.depth.otm.2d}
\alias{otm.depth.otm.2d}
\title{2D Semi-discrete Optimal Transport Depth}
\usage{
\method{otm.depth}{otm.2d}(object, Q, use.geo = FALSE, ...)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}

\item{use.geo}{logical indicating if the geometric method should be used to compute the ranks.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a vector containing the depths of the data.
}
\description{
The 2D implementation of \code{otm.depth}.
}
\keyword{internal}
