% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{otm.depth.OTM_2D}
\alias{otm.depth.OTM_2D}
\title{2D Semi-discrete Optimal Transport Depth}
\usage{
\method{otm.depth}{OTM_2D}(object, Q)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}
}
\value{
a vector containing the depths of the data.
}
\description{
Compute the MK depth relative to the 2D standard uniform distribution.
}
