% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{otm.depth}
\alias{otm.depth}
\title{Semi-discrete Optimal Transport Depth}
\usage{
otm.depth(object, Q, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}

\item{\dots}{additional arguments, currently without effect.}

\item{use.geo}{logical indicating if the geometric method should be used to compute the ranks.}
}
\value{
a vector containing the depths of the data.
}
\description{
\code{otm.depth} computes the optimal transport depth relative to the multivariate uniform distribution \eqn{U[0, 1]^d}.
}
\keyword{multivariate}
