% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{otm.depth}
\alias{otm.depth}
\title{Semi-discrete Optimal Transport Depth}
\usage{
otm.depth(object, Q, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}
}
\value{
a vector containing the depths of the data.
}
\description{
Compute the optimal transport depth relative to the multivariate uniform distribution U[0, 1]^d.
}
