% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.test.R
\name{otm.gof.test}
\alias{otm.gof.test}
\title{2D Goodness-of-test Statistic}
\usage{
otm.gof.test(X, Y, mc = 1000, rank = "center", epsilon = 0.001,
  maxit = 100, verbose = F, na.rm = F)
}
\arguments{
\item{X}{input data matrix, of size \eqn{n} by \eqn{2}.}

\item{Y}{input data matrix, of size \eqn{m} by \eqn{2}.}

\item{mc}{number of quasi-Monte-Carlo samples used to evaluate the test statistic.}

\item{rank}{choose the method for assigning ranks to the data points. Can be "\code{max}", "\code{min}", or "\code{center}".}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating wether to display optimization messages.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\description{
Computes the 2D goodness-of-test statistic using rank defined through the semi-discrete optimal transport map.
}
\keyword{htest,}
\keyword{multivariate}
