% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{otm.potential.OTM.2D}
\alias{otm.potential.OTM.2D}
\title{2D Semi-discrete Optimal Transport Potential}
\usage{
\method{otm.potential}{OTM.2D}(object, Q)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}
}
\value{
a matrix containing the potentials of the queries.
}
\description{
Compute the 2D Alexandrov potential.
}
