% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{otm.potential.otm.2d}
\alias{otm.potential.otm.2d}
\title{2D Semi-discrete Optimal Transport Potential}
\usage{
\method{otm.potential}{otm.2d}(object, Q)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}
}
\value{
a matrix containing the potentials of the queries.
}
\description{
The 2D implementation of \code{otm.potential}.
}
\keyword{internal}
