% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{otm.potential}
\alias{otm.potential}
\title{Semi-discrete Optimal Transport Potential}
\usage{
otm.potential(object, Q, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a matrix containing the potentials of the queries.
}
\description{
\code{otm.potential} computes the Alexandrov potential function.
}
\keyword{multivariate}
