% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{otm.potential}
\alias{otm.potential}
\title{Semi-discrete optimal transport dual potential}
\usage{
otm.potential(object, X, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{X}{a \eqn{n} by \eqn{2} numeric matrix.}
}
\value{
a matrix containing dual potentials of the data.
}
\description{
Compute the Legendre-Fenchel dual of the Alexandrov potential
Semi-discrete optimal transport dual potential.
}
