% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{otm.quantile.otm.2d}
\alias{otm.quantile.otm.2d}
\title{2D Semi-discrete Optimal Transport queryuantile}
\usage{
\method{otm.quantile}{otm.2d}(object, query, scale = TRUE, ...)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point.}

\item{scale}{logical indicating if the queries should be scaled.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a list containing the quantiles of the queries, as well as the indices of the corresponding cells.
}
\description{
The 2D implementation for \code{otm.quantile}.
}
\keyword{internal}
