% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{otm.quantile.OTM.2D}
\alias{otm.quantile.OTM.2D}
\title{2D Semi-discrete Optimal Transport Quantile}
\usage{
\method{otm.quantile}{OTM.2D}(object, Q)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}
}
\value{
a list containing the quantiles of the data, and the indices of the corresponding cells.
}
\description{
Compute the 2D optimal transport quantiles.
}
