% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{otm.quantile}
\alias{otm.quantile}
\title{Semi-discrete Optimal Transport Quantile}
\usage{
otm.quantile(object, Q, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a list containing the quantiles of the queries, as well as the indices of the corresponding cells.
}
\description{
\code{otm.quantile} computes the optimal transport quantiles.
}
\keyword{multivariate}
