% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{otm.rank.OTM.2D}
\alias{otm.rank.OTM.2D}
\title{2D Semi-discrete Optimal Transport Rank}
\usage{
\method{otm.rank}{OTM.2D}(object, Q, use.geo = FALSE)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}

\item{use.geo}{logical indicating if the geometric method should be used to compute the ranks.}
}
\value{
a matrix containing the ranks of the data.
}
\description{
Compute the 2D ranks based on optimal transport.
}
