% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{otm.rank}
\alias{otm.rank}
\title{Semi-discrete Optimal Transport Rank}
\usage{
otm.rank(object, X, h, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{X}{a \eqn{n} by \eqn{2} numeric matrix.}

\item{h}{step size for the numerical differentiation.}
}
\value{
a matrix containing the rank of the data.
}
\description{
Compute the rank function based on optimal transport
Semi-discrete optimal transport rank.
}
