% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{otm.rank}
\alias{otm.rank}
\title{Semi-discrete Optimal Transport Rank}
\usage{
otm.rank(object, Q, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{Q}{a numeric matrix where each row represents a query point.}

\item{\dots}{additional arguments.}
}
\value{
a list containing the ranks of the data and the corresponding convex conjugate potential values.
If \code{use.geo = TRUE} then the convex conjugates will not be computed (\code{NA}s will be returned).
}
\description{
\code{otm.rank} computes the optimal transport ranks.
}
\keyword{multivariate}
