% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OTM_2D.R
\name{OTM_2D}
\alias{OTM_2D}
\title{2D Semi-discrete Optimal Transport Map}
\usage{
OTM_2D(data, epsilon = 0.001, maxit = 100, verbose = F, na.rm = F)
}
\arguments{
\item{data}{input coordinate matrix, of size \eqn{n} by \eqn{2}.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating wether to display optimization messages.}

\item{na.rm}{logical indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
a list describing the resulting optimal transport map.
}
\description{
Computes the semi-discrete optimal transport map from a uniform measure on [0, 1]^2 to a given data set.
}
\keyword{graphs}
\keyword{optimize,}
