% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{pc.perm}
\alias{pc.perm}
\title{Generate \eqn{m} Distinct Permutations of \eqn{n} Elements}
\usage{
pc.perm(n, m)
}
\arguments{
\item{n}{number of elements to permute.}

\item{m}{number of permutations to generate.}
}
\value{
an \eqn{m} by \eqn{n} matrix of indices where each row represents a permutation. 
Rows beyond the \eqn{n!}-th one will be filled with 0s.
}
\description{
\code{pc.perm} generates the first \eqn{m} distinct permutations of \eqn{n} elements with the Plain Changes, 
a.k.a. Steinhaus-Johnson-Trotter, algorithm.
}
\references{
Donald E. Knuth. \emph{The Art of Computer Programming: Combinatorial Algorithms, Part 1}. 1st. 
Vol. 4A. Art of Computer Programming. Addison-Wesley Professional, 2011. Sec. 7.2.1.2. 
isbn: 978-0-201-03804-0.
}
\keyword{utilities}
