% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otm.R
\name{plot.tos.2d}
\alias{plot.tos.2d}
\title{Plotting 2D Semi-discrete Optimal Transport Map}
\usage{
\method{plot}{tos.2d}(
  x,
  which = "Both",
  col.data = "dodgerblue3",
  col.center = "darkorange1",
  col.edge = "black",
  draw.data = T,
  draw.center = T,
  draw.map = F,
  xlim = c(0, 1),
  ylim = c(0, 1),
  xlab = expression("x"),
  ylab = expression("y"),
  pch = 20,
  ...
)
}
\arguments{
\item{x}{a fitted \code{tos.2d} object.}

\item{which}{specify which graph(s) to plot. Can be \code{None}, \code{RVD}, \code{RDT}, or \code{Both}.}

\item{col.data}{color of the data points.}

\item{col.center}{color of the Voronoi centroids.}

\item{col.edge}{color of the edges in plotting RVD and RDT.}

\item{draw.data}{logical indicating if the data points should be plotted.}

\item{draw.center}{logical indicating if the centroids should be plotted.}

\item{draw.map}{logical indicating if dashed lines should be added to show mapping between the data and the Voronoi cells.}

\item{xlim}{the x limits of the plot.}

\item{ylim}{the y limits of the plot.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{pch}{a vector of plotting characters or symbols.}

\item{\dots}{other graphical parameters to plot.}
}
\description{
\cdoe{plot.tos.2d} plots the restricted Voronoi diagram (RVD) and the restricted Delaunay triangulation (RDT) of a given
2D semi-discrete optimal transport map.
}
\keyword{hplot}
