% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otm.R
\name{plot.tos.3d}
\alias{plot.tos.3d}
\title{Plotting the 3D Semi-discrete Optimal Transport Map}
\usage{
\method{plot}{tos.3d}(
  x,
  inter = interactive(),
  which = "Both",
  col.data = "dodgerblue3",
  col.center = "darkorange1",
  col.rvd = "black",
  col.rdt = "firebrick3",
  draw.data = TRUE,
  draw.center = TRUE,
  draw.map = FALSE,
  draw.id = NULL,
  xlim = c(0, 1),
  ylim = c(0, 1),
  zlim = c(0, 1),
  xlab = expression("x"),
  ylab = expression("y"),
  zlab = expression("z"),
  size = 10,
  phi = 45,
  theta = 45,
  pch = 20,
  ...
)
}
\arguments{
\item{x}{a fitted \code{tos.3d} object.}

\item{inter}{logical indicating if the plot should be interactive.}

\item{which}{specify which graph(s) to plot. Can be \code{None}, \code{RVD}, \code{RDT}, or \code{Both}.}

\item{col.data}{color of the data points.}

\item{col.center}{color of the Voronoi centroids.}

\item{col.rvd}{color of the edges in RVD plot.}

\item{col.rdt}{color of the edges in RDT plot.}

\item{draw.data}{logical indicating if the data points should be plotted.}

\item{draw.center}{logical indicating if the centroids should be plotted.}

\item{draw.map}{logical indicating if lines should be added to show mapping between the data and the Voronoi cells.}

\item{draw.id}{vector of indices specifying which data points (and the corresponding cells in RVD and RDT) will be plotted. 
Set to \code{NULL} to plot all points. If it is not \code{NULL} and RDT is to be drawn, 
\code{plot.tos.3d} will plot all tetrahedra that have points in \code{draw.id} as one of its vertices.}

\item{xlim}{the x limits of the plot.}

\item{ylim}{the y limits of the plot.}

\item{zlim}{the z limits of the plot.}

\item{xlab}{a label for the x axis.}

\item{ylab}{a label for the y axis.}

\item{zlab}{a label for the z axis.}

\item{size}{size of the plotted points in the interactive plots.}

\item{theta, phi}{the angles defining the viewing direction for non-interactive plots. 
\code{theta} gives the azimuthal direction and \code{phi} the colatitude.}

\item{pch}{a vector of plotting characters or symbols.}

\item{\dots}{other graphical parameters to plot.}
}
\description{
\code{plot.tos.3d} plots the restricted Voronoi diagram (RVD) and the restricted Delaunay triangulation (RDT) of a given
3D semi-discrete optimal transport map.
}
\examples{
# generate some data
X = c(2.3, 0.8, 1.2, 0.6, 0.2, 0.5, -0.7, -0.6, -0.2, 3.2)
Y = c(1.5, -0.8, 0.6, -2.1, 1.3, 1.3, -1.1, 1.0, -0.2, 1.0)
Z = c(0.8, -0.3, -1.3, -1.8, 1.2, -0.7, -0.8, 1.3, -0.9, 2.0)

# compute the optimal transport map from U[0, 1]^3 to the data
XYZ.OTM = tos.fit(cbind(X, Y, Z))

# plot the (non-interactive) RVD
plot(XYZ.OTM, interactive = FALSE, which = "RVD", draw.map = TRUE)
}
\keyword{hplot}
