% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scaling.min.max}
\alias{scaling.min.max}
\title{Min-max scaling}
\usage{
scaling.min.max(data, min = 0, max = 1)
}
\arguments{
\item{data}{a numeric matrix for the input data.}

\item{min}{desired minimum of the transformed data.}

\item{max}{desired maximum of the transformed data.}
}
\value{
a matrix containing the scaled data.
}
\description{
\code{scaling.min.max} rescales the range of a data set to [min, max].
}
\keyword{utilities}
