% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{scaling.min.max}
\alias{scaling.min.max}
\title{Min-max scaling}
\usage{
scaling.min.max(data, min = 0, max = 1)
}
\arguments{
\item{data}{input data matrix, of size \eqn{n} by \eqn{d}.}

\item{min}{desired minimum of transformed data.}

\item{max}{desired maximum of transformed data.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
a matrix containing the scaled data.
}
\description{
Rescale the range of data to [min, max].
}
\keyword{utilities}
