% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testOTM-package.R
\docType{package}
\name{testOTM-package}
\alias{testOTM}
\alias{testOTM-package}
\title{testOTM: Optimal transport with Geogram}
\description{
Optimal transport in R using the geogram library.
}
\details{
Computes optimal transportation map with Geogram.
}
\references{
Geogram, \url{http://alice.loria.fr/software/geogram/doc/html/index.html}.
}
\author{
\strong{Maintainer}: Peng Xu \email{px2132@columbia.edu}

}
\keyword{internal}
