% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testOTM-package.R
\docType{package}
\encoding{UTF-8}
\name{testOTM}
\alias{testOTM}
\alias{testOTM-package}
\title{testOTM: Multivariate Ranks and Quantiles by Optimal Transportation}
\description{
Multivariate ranks and quantiles based on the theory of optimal transportation, following the definitions by Chernozhukov et al. (2017) <doi:10.1214/16-AOS1450>. This package also provides applications of these notions in nonparametric testing, which are suggested by Ghosal and Sen (2019) <arXiv:1905.05340>. The optimal transport maps are computed by the Geogram library, using methods mentioned in Lévy and Schwindt (2018) <doi:10.1016/j.cag.2018.01.009>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Francis-Hsu/testOTM}
  \item Report bugs at \url{https://github.com/Francis-Hsu/testOTM/issues}
}

}
\author{
\strong{Maintainer}: Peng Xu \email{px2132@columbia.edu}

Other contributors:
\itemize{
  \item Bruno Lévy (Geogram) [contributor, copyright holder]
}

}
\keyword{internal}
