% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{tos.depth}
\alias{tos.depth}
\title{Semi-discrete Optimal Transport Depth}
\usage{
tos.depth(object, query, scale = TRUE, rank.data = "uniform", ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point.}

\item{scale}{logical indicating if the queries should be scaled.}

\item{rank.data}{choose the method for assigning ranks to the data points.}

\item{\dots}{additional arguments.}
}
\value{
a vector containing depths of the queries.
}
\description{
\code{tos.depth} computes the optimal transport depth relative to the multivariate uniform distribution \eqn{U[0, 1]^d}.
}
\references{
Victor Chernozhukov, Alfred Galichon, Marc Hallin, and Marc Henry. 2017.
\emph{Monge–Kantorovich Depth, Quantiles, Ranks and Signs}.
The Annals of Statistics 45 (1): 223–56.
\url{https://doi.org/10.1214/16-AOS1450}.
}
\keyword{multivariate}
