% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth.R
\name{tos.depth.tos.2d}
\alias{tos.depth.tos.2d}
\title{2D Semi-discrete Optimal Transport Depth}
\usage{
\method{tos.depth}{tos.2d}(object, query, scale = TRUE,
  rank.data = "uniform", ...)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point.}

\item{rank.data}{choose the method for assigning ranks to the data points.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a vector containing the depths of the data.
}
\description{
The 2D implementation of \code{tos.depth}.
}
\keyword{internal}
