% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otm.R
\name{tos.fit}
\alias{tos.fit}
\title{Uniform Semi-discrete Optimal Transport Map}
\usage{
tos.fit(
  data,
  scale = c(0.05, 0.95),
  na.rm = F,
  epsilon = 1e-06,
  maxit = 100,
  verbose = F
)
}
\arguments{
\item{data}{a numeric matrix for the input data, of size \eqn{n} by \eqn{d}.}

\item{scale}{a numeric vector indicating the minimum and maximum of the scaled data. Set to \code{NULL} to skip scaling.}

\item{na.rm}{logical indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{epsilon}{convergence threshold for optimization.}

\item{maxit}{max number of iterations before termination.}

\item{verbose}{logical indicating whether to display optimization messages.}
}
\value{
\code{tos.fit} returns an object of class "\code{tos.2d}" or "\code{tos.3d}", depending on the dimension of input data.
An object of class "\code{tos}" is a list describing the resulting optimal transport map.
}
\description{
\code{tos.fit} computes the semi-discrete optimal transport map from the \eqn{U[0, 1]^d} measure to the input data set.
}
\details{
Input data needs be scaled within the [0, 1] range for computation. 
Supply \code{scale} to let the function handles the scaling internally.
Locations and scales will be returned to help transforming the results back to their original range.
}
\keyword{graphs}
\keyword{optimize,}
