% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{tos.potential}
\alias{tos.potential}
\title{Semi-discrete Optimal Transport Potential}
\usage{
tos.potential(object, query, scale = TRUE, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point.}

\item{scale}{logical indicating if the queries should be scaled.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a matrix containing the potentials of the queries.
}
\description{
\code{tos.potential} computes the Alexandrov potential function.
}
\references{
Xianfeng Gu, Feng Luo, Jian Sun, and Shing-Tung Yau. 2015.
\emph{Variational Principles for Minkowski Type Problems, Discrete Optimal Transport, and Discrete Monge-Ampere Equations}.
Asian Journal of Mathematics 20 (January).
\url{https://doi.org/10.4310/AJM.2016.v20.n2.a7}.
}
\keyword{multivariate}
