% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potential.R
\name{tos.potential.tos.2d}
\alias{tos.potential.tos.2d}
\title{2D Semi-discrete Optimal Transport Potential}
\usage{
\method{tos.potential}{tos.2d}(object, query, scale = TRUE, ...)
}
\arguments{
\item{object}{a fitted 2D optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point.}

\item{scale}{logical indicating if the queries should be scaled.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a matrix containing the potentials of the queries.
}
\description{
The 2D implementation of \code{tos.potential}.
}
\keyword{internal}
