% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{tos.quantile}
\alias{tos.quantile}
\title{Semi-discrete Optimal Transport Quantiles}
\usage{
tos.quantile(object, query, scale = TRUE, ...)
}
\arguments{
\item{object}{a fitted optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point.}

\item{scale}{logical indicating if the queries should be scaled.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a list containing the quantiles of the queries and the indices of the corresponding cells.
}
\description{
\code{tos.quantile} computes the optimal transport quantiles.
}
\references{
Promit Ghosal and Bodhisattva Sen. 2019. 
\emph{Multivariate Ranks and Quantiles Using Optimal Transportation and Applications to Goodness-of-Fit Testing}.
\url{http://arxiv.org/abs/1905.05340}.
}
\keyword{multivariate}
