% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{tos.quantile.tos.3d}
\alias{tos.quantile.tos.3d}
\title{3D Semi-discrete Optimal Transport Quantiles}
\usage{
\method{tos.quantile}{tos.3d}(object, query, scale = TRUE, ...)
}
\arguments{
\item{object}{a fitted 3D optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point.}

\item{scale}{logical indicating if the queries should be scaled.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a list containing the quantiles of the queries, as well as the indices of the corresponding cells.
}
\description{
The 3D implementation for \code{tos.quantile}.
}
\keyword{internal}
