% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{tos.rank.tos.3d}
\alias{tos.rank.tos.3d}
\title{3D Semi-discrete Optimal Transport Rank}
\usage{
\method{tos.rank}{tos.3d}(object, query, scale = TRUE, rank.data = "center", ...)
}
\arguments{
\item{object}{a fitted 3D optimal transport map object.}

\item{query}{a numeric matrix where each row represents a query point.}

\item{scale}{logical indicating if the queries should be scaled.}

\item{rank.data}{choose the method for assigning ranks to the data points. 
Can be \code{max}, \code{min}, or \code{center}.}

\item{\dots}{additional arguments, currently without effect.}
}
\value{
a list containing the ranks of the data and the corresponding convex conjugate potential values.
}
\description{
The 3D implementation of \code{tos.rank}.
}
\keyword{internal}
