% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unscaling.min.max}
\alias{unscaling.min.max}
\title{Undo min-max scaling}
\usage{
unscaling.min.max(data, center, scale)
}
\arguments{
\item{data}{a numeric matrix for the input data.}

\item{center}{a numeric vector containing the locations for transforming the data.}

\item{scale}{a numeric vector containing the scales for transforming the data.}
}
\value{
a matrix containing the unscaled data.
}
\description{
Undo min-max scaling
}
\examples{
# generate some data
X = c(1.5, 3.2, 2.6, 0.3, 5.1)
Y = c(3.3, 6.4, 1.8, 0.7, 2.2)

# scale (X, Y) into [0, 1] range
Z = scaling.min.max(cbind(X, Y), 0, 1)

# undo the min-max scaling
unscaling.min.max(Z, attr(Z, "scaled:center"), attr(Z, "scaled:scale"))
}
\keyword{utilities}
